refactor: renamed Layout and Navbar components; remove unused Sidebar component
