+++ src/components/Layout/Layout.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Navbar from './Navbar';
-import Sidebar from './Sidebar';
+import Navbar from '../Navbar/Navbar';
+import Sidebar from '../Sidebar/Sidebar';
 import styled from 'styled-components';
 
 interface LayoutProps {
